backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
statsmodels
yfinance
pydantic
supabase
python-dotenv
email-validator
scikit-learn>=1.6.1